Change Makefile tag to latest
Currently the Makefile builds the '1.0.0-alpha1' container tag (by
default). Given that our release process expects this version to be fed
through gitlab-ci, this tag will only be built "locally".
It seems reasonable then to have the value of that tag be latest. It
also frees us from having to update this tag for eveery release.

Signed-off-by: Renaud Gaubert <rgaubert@nvidia.com>