Move appending of 'toolkit' outside of subcommands
Previously, each subcommand was appending the path part 'toolkit'  to
the end of the DESTINATION paramater passed into it. This is an odd
abstraction to have, as it makes more sense to have the directory passed
in be the location of the toolkit itself (and not the parent folder of
it). This patch makes sure that this is the case and fixes up the
documentation around it in the usage strings.