Updated logic to get PID of dockerd for signaling via SIGHUP
Unfortunately, the previous method used to signal dockerd will not work
for because docker typically gets its socketfile from systemd. As such,
with the previous approach, we end up signaling systemd instead of
docker.

The problem is that systemd is the one who calls bind() and listen() on
the socket. And when it forks dockerd, it just has dockerd keep the fds
for those open so that it can call accept(), send(), and recv() on that
already bound socket. As such, when the previous code asked for the
process “listening” on the socket, it got the pid of systemd and not
dockerd.

So instead of using unix.GetsockoptUcred() against the listener, this
commit now contacts dockerd directly (by issuing an HTTP /info call
against it) and pulling the PID from the ancillary data embedded in the
response.

Signed-off-by: Kevin Klues <kklues@nvidia.com>