Refactor CI for public repo
This change splits the build and test steps to be run for all build
targets in separate steps. This also adds tests for non-ubuntu18
targets, but these are currently skipped.

A deploy step for pushing images to the CI repository is also added
to mirror the behavior of the internal repo.

Signed-off-by: Evan Lezar <elezar@nvidia.com>