Migrate toolkit.sh to go
This change introduces a toolkit command (with install and delete subcommands)
that installs / removes the NVIDIA container toolkit.

'install' copies the components of the container toolkit to the specified
location. This includes creating symlinks to the correct library and
wrappers for the executables.

'delete' removes the toolkit installation specified directory

Signed-off-by: Evan Lezar <elezar@nvidia.com>