Write and update nvidia-runtime-cli config as toml
This change updates the TOML config for the nvidia-runtime-cli
directly instead of applying string replacements. This means that

nvidia-runtime-cli.root is explicitly set to /run/nvidia/driver
nvidia-runtime-cli.ldconfig is set to @/run/nvidia/driver/sbin/ldconfig[.real]
nvidia-runtime-cli.path is set to the path for the installed executable

Other non-commented options are taken from the source config.

*Note:* that this changes the behaviour in the following ways:
* The ldconfig setting is changed to @/run/nvidia/driver/sbin/ldconfig on
  non-ubuntu systems. This was previously left as @/sbin/ldconfig
* The path value is explicitly set

Signed-off-by: Evan Lezar <elezar@nvidia.com>