Explicitly set the version on containerd config files
In the case where the config file is empty and the version is inferred
from the containerd version, the v2 config file ends up being invalid
if no version is added. This change explictly sets the version of the
generated config.

A test for an empty config file is also added.

Signed-off-by: Evan Lezar <elezar@nvidia.com>