Fix setting of default runtime for v1 configs earlier than containerd 1.3
This change fixes a bug that resulted in the default runtime always being set.
A containerdVersion type is introduced that checks validity and adds an atLeast
method that makes the intent of the comparisons clearer.

Unit tests covering the setting of the default runtimes for v1 configs are also
added.

Signed-off-by: Evan Lezar <elezar@nvidia.com>