Make set-as-default command line flag hidden
The `true` default value for the `--set-as-default` command line flag means that
this option cannot be disabled from the command line. To ensure that this is not
confusing to the user, the flag is set as hidden to ensure that it does not show
up in the help.

This change also removes the default values for false-by-default boolean flags used in the CLIs.
The behaviour for these values (--no-daemon and --no-signal-containerd) will not change,
but the change keeps the code consistent.

Signed-off-by: Evan Lezar <elezar@nvidia.com>