Set LD_LIBRARY_PATH for experimental runtime
The NVML-based discovery phase of the experimental NVIDIA Container Runtime
requires libnvidia-ml.so. This change ensures that LD_LIBRARY_PATH is set
so that this library can be found.

The common library search code is also factored out to allow for it to be
used when searching for the container library.

Signed-off-by: Evan Lezar <elezar@nvidia.com>