Add --runtime-type flag to contianerd config.
This change adds a --runtime-type (CONTAINERD_RUNTIME_TYPE) flag to the CLI
for configuring containerd. This allows the runtime type to be set explicitly
instead of relying on heuristics based on the config version.

The default value for the flag is io.containerd.runc.v2 which is the default
since containerd 1.4 and is supported in older verisons.

Signed-off-by: Evan Lezar <elezar@nvidia.com>