Add --use-legacy-config option to avoid containerd version detection
In order to determine the default config version, the containerd version
is automatically detected. This is brittle as it relies on version
comparisons that need to support any installed containerd version. In
certain cases (such as with an ubuntu package), the version reported by
containerd is not considered a valid semantic version by the package used.

This change removes the need for version comparisons by adding a command line
argument and associated environment variable to switch to the use of the
version 1 config files. This should only be needed for containerd versions
prior to v1.3.

Signed-off-by: Evan Lezar <elezar@nvidia.com>